Time: 1 ms (85.56%), Space: 42.6 MB (31.07%) - LeetHub

+++ 0029-divide-two-integers/0029-divide-two-integers.java
@@ -6,7 +6,7 @@ class Solution {
         }
         
         // Use long to avoid overflow
-        long result = (long) dividend / (long) divisor;
+        long result =  dividend /  divisor;
         
         // Check bounds
         if (result > Integer.MAX_VALUE) return Integer.MAX_VALUE;
